requirements.txt
numpy
streamlit
Pillow
keras
tensorflow
matplotlib
requests